// client/src/types/enums.tsx
export enum MediaActions {
	SET_ALL_MEDIA = 'media/set_all',
	SET_CURRENT_MEDIA = 'media/set_current',

	SET_ALL_SERIES = 'series/set_all',
	SET_CURRENT_SERIES = 'series/set_current',
	SET_SERIES_MOVIES_ALL = 'series/movies/set_all',

	SET_ALL_SEASONS = 'seasons/set_all',
	SET_CURRENT_SEASON = 'seasons/set_current',
	SET_SEASON_EPISODES_ALL = 'seasons/episodes/set_all',
}

export enum Page {
	MOVIES = 'movies',
	SERIES = 'series',
	ALL = 'all',
}